refactor(cli): move pure filename/timestamp helpers into utils
